Avoid case of illegal ponder move [bug]
When the main line consists of a game-terminating move (such as a
draw), and a research is active and a ponderhit or timeout occurs,
the first move of the research will be mistaken for a candidate
pondermove becuase it starts in slot 1. But this is always an illegal
ponder move, as the side to move doesn't even match.

Solution is to insert 0-valued terminators between the temporary
PV under construction and the current PV. This way in research a PV
never starts before slot 2 and can't be confused for a ponder move.
In the above scenario, the pondermove would be 0, which is ok.

The added benefit is for debugging when dumping the self->pv list:
The zeroes are clear separators between PV segments.