Workaround cutechess ponder problem on conversion to KK, KNK or KBK [usability]
Still move confusion can arise. See also
http://talkchess.com/forum/viewtopic.php?p=648257#648257

Shredder simply swallows 'bestmove' in case of `ucinewgame' or a
second `go', but this is difficult to do in Floyd. Stockfish hangs.

Changes are:
- evaluate() gives 0 in KK, KNK and KB...K for like-bishops
- Don't allow 0 as heuristic evaluation otherwise
- Stop PV when evaluate gives 0, so that such conversion can not become a pondermove

Better solution is to use internal node recognition also in scout,
but that is a too big change for now.