Workaround for '0readyok' problem in python-chess [bug,v0.8.1]
Python-chess can send 'isready' while the engine is calculating.
If Floyd is in the middle of composing an info string, the 'readyok'
can then appear mangled in the info line, such as '0readyok'.
Python-chess than hangs on an exception, for example

  ValueError: invalid literal for int() with base 10: '13487readyok'

This workaround is to add an additional newline befor 'readyok'.
This is a bit ugly. I think it is still safe, because the stdio is
thread-safe otherwise, and the info string elements are all written
with printf commands containing both key and value together.  The
info line will still be broken up, and the tail hopefully swallowed
by the engine.

A proper solution would be to write the info line in a local buffer
first before sending it to stdout with a single puts(). This should
be considered for the v0.9 revision of Floyd.