Alternative solution for '0readyok' problem [qa]
The temporarily solution in v0.8.1 simply inserted a newline before
'readyok'. This solves only half of the problem as it will still
lose information when split in the middle.

If we were only on POSIX it would be sufficient to use flockfile()
and funlockfile() in uciSearchInfo(). However floyd must also work
in windows and there these functions are not available.  Explicit
locking with mutexes gets very messy. You would have to do it
everywhere.

Therefore the second best solution is to prepare the info lines in
a local buffer and write it out in one stdio call at the end.  The
assumption is that stdio calls are atomic in a multi-threaded
environment. This seems to be the case in C11, POSIX and windows,
although I haven't tried to disprove it by testing.
http://stackoverflow.com/questions/22539282/is-msvcrts-implementation-of-fprintf-thread-safe

This solution takes more lines than the previous solution or the
POSIX solution, but it is not real bad either. Essentially the
implementaiton of listPrintf in cplus.c. On the application side
the number of lines is not really different.