Accept lower-case squashing of "setoption" names [uci,bug]
This mainly affects changing the hash size under PicoChess/python-chess.

The UCI spec says this:
"""
* setoption name <id> [value <x>]
[...]
The name and value of the option in <id> should not be case sensitive and can inlude [sic] spaces.
"""

[Note MvK: It is unclear to me what is meant with the "value" in
"<id>". Does it refer to <x>?  If so, how can that even work if
values can be path names?]

This is more of a workaround, but I don't want to spend potentially
a lot of lines to do it completely "right". Simply squashing to
lowercase ourselves before sscanf() is problematic for two reasons:

1. The line buffer contents after " value " must not be affected.
So you would get something ugly to stop there as well.
        for (i=0; line[i]&&strncmp(line+i," value ",7); i++)
                line[i] = tolower(line[i]);
This is already not worth it.

2. Modifying line buffer in place is furthermore problemetic because
the same buffer is also used to echo back input in error messages
in debug mode.  If an option is not recognized, the original spelling
should be used there.

Therefore, recognizing both all-lowercase and the original
capitalization of "uci" command output is a compromise. Lets see
how long it lasts.